requirements.txt
fastapi
tortoise-orm
asyncmy
pydantic
httpx
pytest
pytest-asyncio
aerich
python-dotenv
uvicorn